report-generation/requirements.txt
Flask~=2.0.0
Pygments~=2.10.0
Werkzeug~=2.0.1
humanize~=3.11.0
ipython~=7.27.0
jmespath~=0.10.0
docutils~=0.17.1
Markdown~=3.3.4
Sphinx~=4.1.2
numpy~=1.21.2
flask-socketio~=5.1.1
jsonschema~=3.2.0
paramiko~=2.8.0
pymongo~=4.0.1
Jinja2~=3.0.3
ansi2html~=1.7.0
Flask-Markdown~=0.3
docker
python-socketio[Client]